LottoResult에 LottoTicket 가지는 걸로 변경
